added convergence testing for fdtd and pstd
